Created new Splash screen
Added a new splash screen with a new design and functionality. The splash screen now displays the app logo and loading text, then navigates to the login screen after a 2-second delay.

Refactored the main.dart file to remove the old splash screen implementation and set the initial route to the new splash screen.

Updated the login screen layout to be more responsive and centered.